+++ src/K4os.KnownTypes/KnownTypesRegistry.cs
@@ -9,7 +9,7 @@ namespace K4os.KnownTypes;
 /// Know types registry storing list of well-known type aliases to, for example,
 /// avoid usage of fully qualified type names.
 /// </summary>
-public class KnownTypesRegistry
+public class KnownTypesRegistry: IKnownTypesRegistry, IKnownTypesResolver
 {
     /// <summary>The default <see cref="KnownTypesRegistry"/></summary>
     public static readonly KnownTypesRegistry Default = new();
@@ -20,8 +20,23 @@ public class KnownTypesRegistry
     private volatile IReadOnlyDictionary<string, Type>? _frozenNameToType;
     private volatile IReadOnlyDictionary<Type, string>? _frozenTypeToName;
     private volatile IReadOnlyList<Type>? _frozenKnownTypes;
+    private readonly IKnownTypeAliasExtractor _extractor;
 
     private readonly object _lock = new();
+
+    /// <summary>
+    /// Creates new instance of <see cref="KnownTypesRegistry"/> with
+    /// default <see cref="IKnownTypeAliasExtractor"/>.
+    /// </summary>
+    public KnownTypesRegistry(): this(DefaultKnownTypeAliasExtractor.Instance) { }
+
+    /// <summary>
+    /// Creates new instance of <see cref="KnownTypesRegistry"/> with given
+    /// <see cref="IKnownTypeAliasExtractor"/>.
+    /// </summary>
+    /// <param name="extractor">Alias extractor.</param>
+    public KnownTypesRegistry(IKnownTypeAliasExtractor extractor) =>
+        _extractor = extractor;
 
     private IReadOnlyDictionary<K, V> Freeze<K, V>(Dictionary<K, V> source) where K: notnull
     {
@@ -31,21 +46,21 @@ public class KnownTypesRegistry
     private IReadOnlyDictionary<string, Type> FrozenNameToType() =>
         // ReSharper disable once NonAtomicCompoundOperator
         _frozenNameToType ??= Freeze(_nameToType);
-    
+
     private IReadOnlyDictionary<Type, string> FrozenTypeToName() =>
         // ReSharper disable once NonAtomicCompoundOperator
         _frozenTypeToName ??= Freeze(_typeToName);
-    
+
     private IReadOnlyList<Type> FrozenKnownTypes() =>
         // ReSharper disable once NonAtomicCompoundOperator
         _frozenKnownTypes ??= FrozenTypeToName().Keys.ToImmutableArray();
 
     private Type? TryResolve(string name) =>
         FrozenNameToType().GetValueOrDefault(name);
-    
+
     private string? TryResolve(Type type) =>
         FrozenTypeToName().GetValueOrDefault(type);
-    
+
     private void TryRegisterImpl(string name, Type type)
     {
         // one type may have many names, but name may point to only one type
@@ -71,83 +86,76 @@ public class KnownTypesRegistry
         _frozenKnownTypes = null;
     }
 
-    /// <summary>Registers known polymorphic type under specified name.
-    /// Type needs to be annotated with <see cref="KnownTypeAliasAttribute"/></summary>
-    /// <typeparam name="T">Annotated known type.</typeparam>
-    public void Register<T>() => Register(typeof(T));
-
-    /// <summary>Registers known polymorphic type under specified name.
-    /// Type needs to be annotated with <see cref="KnownTypeAliasAttribute"/></summary>
-    /// <param name="type">The known polymorphic type.</param>
-    public void Register(Type type) => Register(type.GetTypeInfo());
-
-    /// <summary>Registers known polymorphic type under specified name.
-    /// Type needs to be annotated with <see cref="KnownTypeAliasAttribute"/></summary>
-    /// <param name="typeInfo">The known polymorphic type.</param>
-    public void Register(TypeInfo typeInfo)
+    /// <inheritdoc />
+    public void Register(Type type, string alias)
     {
-        var names = KnownTypeAliasAttribute.EnumerateNames(typeInfo) switch {
-            null or { Length: 0 } => [typeInfo.FullName.ThrowIfNull("TypeInfo.FullName")], 
-            var list => list
-        };
-        var type = typeInfo.AsType();
-
         lock (_lock)
         {
-            foreach (var name in names)
+            TryRegisterImpl(alias, type);
+        }
+    }
+
+    /// <inheritdoc />
+    public void Register(Type type, IEnumerable<string>? aliases)
+    {
+        lock (_lock)
+        {
+            foreach (var name in EnumerateAliases(type, aliases))
                 TryRegisterImpl(name, type);
         }
     }
 
-    /// <summary>Registers all types from assembly containing give type.</summary>
-    /// <typeparam name="T">Hook type.</typeparam>
-    public void RegisterAssembly<T>() => RegisterAssembly(typeof(T));
-
-    /// <summary>Registers all types from assembly containing give type.</summary>
-    /// <param name="hookType">Hook type.</param>
-    public void RegisterAssembly(Type hookType) => RegisterAssembly(hookType.GetTypeInfo());
-
-    /// <summary>Registers all types from assembly containing give type.</summary>
-    /// <param name="hookType">Hook type.</param>
-    public void RegisterAssembly(TypeInfo hookType) => RegisterAssembly(hookType.Assembly);
-
-    /// <summary>Register all types in assembly with <see cref="KnownTypeAliasAttribute"/> annotation.</summary>
-    /// <param name="assembly">Assembly.</param>
-    public void RegisterAssembly(Assembly assembly)
-    {
-        var types = assembly
-            .DefinedTypes
-            .Where(ti => ti.GetCustomAttributes<KnownTypeAliasAttribute>().Any())
-            .ToArray();
-
-        foreach (var type in types)
-            Register(type);
-    }
-
-    /// <summary>Registers known polymorphic type under specified name.</summary>
-    /// <param name="name">The JSON friendly name.</param>
-    /// <param name="type">The known polymorphic type.</param>
-    public void Register(string name, Type type)
+    /// <inheritdoc />
+    public void Register(
+        IEnumerable<Type> types, Func<Type, IEnumerable<string>?> aliases)
     {
         lock (_lock)
         {
-            TryRegisterImpl(name, type);
+            foreach (var type in types)
+            foreach (var name in EnumerateAliases(type, aliases))
+                TryRegisterImpl(name, type);
         }
     }
 
-    /// <summary>Returns type associated with given name.</summary>
-    /// <param name="alias">Type alias.</param>
-    /// <returns>Type associated with alias.</returns>
+    /// <inheritdoc />
+    public void Register(
+        Assembly assembly, Func<Type, bool> predicate, Func<Type, IEnumerable<string>?> aliases)
+    {
+        lock (_lock)
+        {
+            foreach (var type in assembly.GetTypes().Where(predicate))
+            foreach (var name in EnumerateAliases(type, aliases))
+                TryRegisterImpl(name, type);
+        }
+    }
+
+    private static IEnumerable<string> EnumerateAliases(
+        Type type, IEnumerable<string>? aliases) =>
+        aliases ?? throw new ArgumentException(
+            $"Type {type.Name} does not have any known aliases");
+
+    private static IEnumerable<string> EnumerateAliases(
+        Type type, Func<Type, IEnumerable<string>?> aliases) =>
+        EnumerateAliases(type, aliases(type));
+
+    /// <inheritdoc />
+    public void Register(Type type)
+    {
+        Register(type, _extractor.GetAliases(type));
+    }
+
+    /// <inheritdoc />
+    public void RegisterAssembly(Assembly assembly)
+    {
+        Register(assembly, _extractor.AutoRegister, t => _extractor.GetAliases(t));
+    }
+
+    /// <inheritdoc />
     public Type? TryGetType(string alias) => TryResolve(alias);
 
-    /// <summary>Returns first alias associated with given type.</summary>
-    /// <param name="type">Type.</param>
-    /// <returns>First alias associated with type.</returns>
+    /// <inheritdoc />
     public string? TryGetAlias(Type type) => TryResolve(type);
-    
-    /// <summary>
-    /// Lists all known types.
-    /// Please note, this method is relatively slow as list itself is not cached.
-    /// </summary>
+
+    /// <inheritdoc />
     public IReadOnlyList<Type> KnownTypes => FrozenKnownTypes();
 }